Collision improvements
Creating a new branch because something went wrong with the other
(or I just got frastrated with git stuff). This was based on master.
Fixed a bug in the simple collision detection and added a couple more
obstacles to the list to test. There is a bug in the color when drawing
other shapes after a collision. Need to add way for the game to
stop after collision